+++ runtime/binding-ws/src/test/java/io/aklivity/zilla/runtime/binding/ws/internal/streams/server/AdvisoryIT.java
@@ -91,4 +91,13 @@ public class AdvisoryIT
     {
         k3po.finish();
     }
+
+    @Test
+    @Configuration("server.yaml")
+    @Specification({
+        "${net}/client.sent.no.upgrade/client" })
+    public void shouldReceiveClientEndWhenNoUpdate() throws Exception
+    {
+        k3po.finish();
+    }
 }