+++ runtime/engine/src/main/java/io/aklivity/zilla/runtime/engine/config/RouteConfig.java
@@ -19,6 +19,7 @@ import static java.util.Objects.requireNonNull;
 import static java.util.function.Function.identity;
 
 import java.util.List;
+import java.util.function.ToLongFunction;
 import java.util.function.UnaryOperator;
 
 import io.aklivity.zilla.runtime.engine.util.function.LongObjectPredicate;
@@ -27,6 +28,7 @@ public class RouteConfig
 {
     public transient long id;
     public transient LongObjectPredicate<UnaryOperator<String>> authorized;
+    public transient ToLongFunction<String> resolveId;
 
     public final int order;
     public final String exit;